chore: Move form address helper into it's own file
